+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # remeasure CHANGELOG
 
+#### 2.4.1
+* Replace `element-resize-detector` with use of `ResizeObserver` ponyfill
+
 #### 2.4.0
-* Add `es` transpilation in publish, for [`pkg.module`](https://github.com/rollup/rollup/wiki/pkg.module) 
+* Add `es` transpilation in publish, for [`pkg.module`](https://github.com/rollup/rollup/wiki/pkg.module)
 
 #### 2.3.1
 * Higher-order component will extend `PureComponent` if the component being wrapped is also a `PureComponent` (extends `Component` in all other scenarios)